+++ src/main/java/com/app/modules/social/entity/FollowId.java
@@ -1,0 +1,28 @@
+package com.app.modules.social.entity;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Embeddable
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+public class FollowId implements Serializable {
+
+    @Column(name = "follower_id", nullable = false)
+    private UUID followerId;
+
+    @Column(name = "following_id", nullable = false)
+    private UUID followingId;
+}